+++ backend/app/requirements.txt
@@ -33,3 +33,5 @@ asgi-lifespan==1.0.1
 ecdsa
 pyblake2
 base58
+
+aiofiles
